Legal/requirements.txt
streamlit
PyMuPDF
gTTS
python-dotenv
requests
pytesseract
Pillow
